src/Rangle.Implementations/Services/DataProtectionService.cs
using System;
using System.IO;
using System.Security.Cryptography;
using System.Text;
using Rangle.Abstractions.Services;
using BC = BCrypt.Net.BCrypt;

namespace Rangle.Implementations.Services
{
    public class DataProtectionService : IDataProtectionService
    {
        public string Encrypt(string text, string privateKey)
        {
            if (privateKey == null)
            {
                throw new ArgumentNullException(nameof(privateKey));
            }

            if (string.IsNullOrWhiteSpace(text))
            {
                return string.Empty;
            }

            byte[] key = new SHA256Managed().ComputeHash(Encoding.UTF8.GetBytes(privateKey));

            using (Aes aesAlg = Aes.Create())
            {
                using (var encryptor = aesAlg.CreateEncryptor(key, aesAlg.IV))
                {
                    using (var msEncrypt = new MemoryStream())
                    {
                        using (var csEncrypt = new CryptoStream(msEncrypt, encryptor, CryptoStreamMode.Write))
                        using (var swEncrypt = new StreamWriter(csEncrypt))
                        {
                            swEncrypt.Write(text);
                        }

                        var iv = aesAlg.IV;

                        var decryptedContent = msEncrypt.ToArray();

                        var result = new byte[iv.Length + decryptedContent.Length];

                        Buffer.BlockCopy(iv, 0, result, 0, iv.Length);
                        Buffer.BlockCopy(decryptedContent, 0, result, iv.Length, decryptedContent.Length);

                        return Convert.ToBase64String(result);
                    }
                }
            }
        }

        public string Decrypt(string encryptedText, string privateKey)
        {
            if (privateKey == null)
            {
                throw new ArgumentNullException(nameof(privateKey));
            }

            if (string.IsNullOrWhiteSpace(encryptedText))
            {
                return string.Empty;
            }

            var fullCipher = Convert.FromBase64String(encryptedText);

            var iv = new byte[16];
            var cipher = new byte[fullCipher.Length - iv.Length];

            Buffer.BlockCopy(fullCipher, 0, iv, 0, iv.Length);
            Buffer.BlockCopy(fullCipher, iv.Length, cipher, 0, cipher.Length);

            byte[] key = new SHA256Managed().ComputeHash(Encoding.UTF8.GetBytes(privateKey));

            using (Aes aesAlg = Aes.Create())
            {
                using (var decryptor = aesAlg.CreateDecryptor(key, iv))
                {
                    string result = string.Empty;
                    using (var msDecrypt = new MemoryStream(cipher))
                    {
                        using (var csDecrypt = new CryptoStream(msDecrypt, decryptor, CryptoStreamMode.Read))
                        {
                            using (var srDecrypt = new StreamReader(csDecrypt))
                            {
                                try
                                {
                                    result = srDecrypt.ReadToEnd();
                                }
                                catch (CryptographicException e)
                                {
                                    // Ignore
                                }                            
                            }
                        }
                    }

                    return result;
                }
            }
        }

        public string HashPassword(string password)
        {
            if (string.IsNullOrEmpty(password))
            {
                throw new ArgumentNullException(nameof(password));
            }

            return BC.HashPassword(password);
        }

        public bool VerifyPassword(string password, string hash)
        {
            if (string.IsNullOrEmpty(password))
            {
                throw new ArgumentNullException(nameof(password));
            }

            if (string.IsNullOrEmpty(hash))
            {
                throw new ArgumentNullException(nameof(hash));
            }

            return BC.Verify(password, hash);
        }
    }
}
